update import for renamed mixin file
